pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentloop"
version = "0.1.0"
description = "Multi-agent closed-loop orchestration system"
authors = [
    {name = "AgentLoop", email = "dev@agentloop.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.12.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
    "uuid7>=0.1.0",
    "python-dateutil>=2.8.2",
    "greenlet>=3.3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
agentloop = "agentloop.main:cli"

[tool.hatch.version]
path = "agentloop/__init__.py"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
